+++ spec/helpers/work_packages_helper_spec.rb
@@ -123,6 +123,126 @@ describe WorkPackagesHelper do
     end
   end
 
+  describe :work_package_css_classes do
+    it "should always have the work_package class" do
+      helper.work_package_css_classes(stub_work_package).should include("work_package")
+    end
+
+    it "should return the position of the work_package's status" do
+      status = double('status', :is_closed? => false)
+
+      stub_work_package.stub!(:status).and_return(status)
+      status.stub!(:position).and_return(5)
+
+      helper.work_package_css_classes(stub_work_package).should include("status-5")
+    end
+
+    it "should not have a status class if the work_package has none" do
+      helper.work_package_css_classes(stub_work_package).should_not include("status")
+    end
+
+    it "should return the position of the work_package's priority" do
+      priority = double('priority')
+
+      stub_work_package.stub!(:priority).and_return(priority)
+      priority.stub!(:position).and_return(5)
+
+      helper.work_package_css_classes(stub_work_package).should include("priority-5")
+    end
+
+    it "should not have a priority class if the work_package has none" do
+      helper.work_package_css_classes(stub_work_package).should_not include("priority")
+    end
+
+    it "should have a closed class if the work_package is closed" do
+      stub_work_package.stub!(:closed?).and_return(true)
+
+      helper.work_package_css_classes(stub_work_package).should include("closed")
+    end
+
+    it "should not have a closed class if the work_package is not closed" do
+      stub_work_package.stub!(:closed?).and_return(false)
+
+      helper.work_package_css_classes(stub_work_package).should_not include("closed")
+    end
+
+    it "should have an overdue class if the work_package is overdue" do
+      stub_work_package.stub!(:overdue?).and_return(true)
+
+      helper.work_package_css_classes(stub_work_package).should include("overdue")
+    end
+
+    it "should not have an overdue class if the work_package is not overdue" do
+      stub_work_package.stub!(:overdue?).and_return(false)
+
+      helper.work_package_css_classes(stub_work_package).should_not include("overdue")
+    end
+
+    it "should have a child class if the work_package is a child" do
+      stub_work_package.stub!(:child?).and_return(true)
+
+      helper.work_package_css_classes(stub_work_package).should include("child")
+    end
+
+    it "should not have a child class if the work_package is not a child" do
+      stub_work_package.stub!(:child?).and_return(false)
+
+      helper.work_package_css_classes(stub_work_package).should_not include("child")
+    end
+
+    it "should have a parent class if the work_package is a parent" do
+      stub_work_package.stub!(:leaf?).and_return(false)
+
+      helper.work_package_css_classes(stub_work_package).should include("parent")
+    end
+
+    it "should not have a parent class if the work_package is not a parent" do
+      stub_work_package.stub!(:leaf?).and_return(true)
+
+      helper.work_package_css_classes(stub_work_package).should_not include("parent")
+    end
+
+    it "should have a created-by-me class if the work_package is a created by the current user" do
+      stub_user = double('user', :logged? => true, :id => 5)
+      User.stub!(:current).and_return(stub_user)
+      stub_work_package.stub!(:author_id).and_return(5)
+
+      helper.work_package_css_classes(stub_work_package).should include("created-by-me")
+    end
+
+    it "should not have a created-by-me class if the work_package is not created by the current user" do
+      stub_user = double('user', :logged? => true, :id => 5)
+      User.stub!(:current).and_return(stub_user)
+      stub_work_package.stub!(:author_id).and_return(4)
+
+      helper.work_package_css_classes(stub_work_package).should_not include("created-by-me")
+    end
+
+    it "should not have a created-by-me class if the work_package is the current user is not logged in" do
+      helper.work_package_css_classes(stub_work_package).should_not include("created-by-me")
+    end
+
+    it "should have a assigned-to-me class if the work_package is a created by the current user" do
+      stub_user = double('user', :logged? => true, :id => 5)
+      User.stub!(:current).and_return(stub_user)
+      stub_work_package.stub!(:assigned_to_id).and_return(5)
+
+      helper.work_package_css_classes(stub_work_package).should include("assigned-to-me")
+    end
+
+    it "should not have a assigned-to-me class if the work_package is not created by the current user" do
+      stub_user = double('user', :logged? => true, :id => 5)
+      User.stub!(:current).and_return(stub_user)
+      stub_work_package.stub!(:assigned_to_id).and_return(4)
+
+      helper.work_package_css_classes(stub_work_package).should_not include("assigned-to-me")
+    end
+
+    it "should not have a assigned-to-me class if the work_package is the current user is not logged in" do
+      helper.work_package_css_classes(stub_work_package).should_not include("assigned-to-me")
+    end
+  end
+
   describe :work_package_form_estimated_hours_attribute do
     it "should output the estimated hours value with a precision of 2" do
       stub_work_package.estimated_hours = 3