+++ doc/CHANGELOG.md
@@ -25,6 +25,7 @@
 * `#1371` Changing pagination per_page_param does not change page
 * `#1314` Always set last activity timestamp and check session expiry if ttl-setting is enabled
 * `#1414` Remove start & due date requirement from planning elements
+* `#1493` Exporting work packages to pdf returns 406
 
 ## 3.0.0pre8